--- a/dspace-server-webapp/src/main/java/org/dspace/app/rest/HarvesterMetadataController.java
+++ b/dspace-server-webapp/src/main/java/org/dspace/app/rest/HarvesterMetadataController.java
@@ -12,6 +12,7 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.dspace.app.rest.converter.ConverterService;
 import org.dspace.app.rest.link.HalLinkService;
 import org.dspace.app.rest.model.HarvesterMetadataRest;
 import org.dspace.app.rest.model.hateoas.HarvesterMetadataResource;
@@ -38,6 +39,9 @@ public class HarvesterMetadataController {
     @Autowired
     private HalLinkService halLinkService;
 
+    @Autowired
+    private ConverterService converter;
+
     /**
      * GET endpoint that returns all available metadata formats
      * @param request   The request object
@@ -52,9 +56,7 @@ public class HarvesterMetadataController {
         HarvesterMetadataRest data = new HarvesterMetadataRest();
         data.setConfigs(configs);
 
-        HarvesterMetadataResource resource = new HarvesterMetadataResource(data, utils);
-        halLinkService.addLinks(resource);
-
+        HarvesterMetadataResource resource = converter.toResource(data);
         return resource;
     }