--- a/dspace-server-webapp/src/main/java/org/dspace/app/rest/RelationshipTypeRestController.java
+++ b/dspace-server-webapp/src/main/java/org/dspace/app/rest/RelationshipTypeRestController.java
@@ -18,6 +18,7 @@ import org.dspace.app.rest.link.HalLinkService;
 import org.dspace.app.rest.model.RelationshipTypeRest;
 import org.dspace.app.rest.model.RelationshipTypeRestWrapper;
 import org.dspace.app.rest.model.hateoas.RelationshipTypeResourceWrapper;
+import org.dspace.app.rest.projection.Projection;
 import org.dspace.app.rest.utils.ContextUtil;
 import org.dspace.app.rest.utils.Utils;
 import org.dspace.content.EntityType;
@@ -67,7 +68,8 @@ public class RelationshipTypeRestController {
      * @throws SQLException If something goes wrong
      */
     @RequestMapping(method = RequestMethod.GET)
-    public RelationshipTypeResourceWrapper retrieve(@PathVariable Integer id, HttpServletResponse response,
+    public RelationshipTypeResourceWrapper retrieve(@PathVariable Integer id,
+                                                    HttpServletResponse response,
                                                     HttpServletRequest request) throws SQLException {
         Context context = ContextUtil.obtainContext(request);
         EntityType entityType = entityTypeService.find(context, id);
@@ -75,12 +77,13 @@ public class RelationshipTypeRestController {
 
         List<RelationshipTypeRest> relationshipTypeRests = new LinkedList<>();
 
+        Projection projection = utils.obtainProjection();
         for (RelationshipType relationshipType : list) {
-            relationshipTypeRests.add(converter.toRest(relationshipType));
+            relationshipTypeRests.add(converter.toRest(relationshipType, projection));
         }
 
-
         RelationshipTypeRestWrapper relationshipTypeRestWrapper = new RelationshipTypeRestWrapper();
+        relationshipTypeRestWrapper.setProjection(projection);
         relationshipTypeRestWrapper.setEntityTypeId(id);
         relationshipTypeRestWrapper.setEntityTypeLabel(entityType.getLabel());
         relationshipTypeRestWrapper.setRelationshipTypeRestList(relationshipTypeRests);