… dspace-server-webapp/src/main/java/org/dspace/app/rest/RestResourceController.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/RootRestResourceController.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/StatisticsRestController.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/UUIDLookupRestController.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/AInprogressItemConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/AuthorityEntryRestConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/AuthorityRestConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/BitstreamConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/BitstreamFormatConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/BrowseIndexConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/ClaimedTaskConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/CollectionConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/CommunityConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/ConverterService.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DSpaceObjectConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DiscoverConfigurationConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DiscoverFacetResultsConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DiscoverFacetValueConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DiscoverFacetsConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/DiscoverResultConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/EPersonConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/EntityTypeConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/GenericDSpaceObjectConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/GroupConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/HarvestedCollectionConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/IndexableObjectConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/ItemConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/MetadataConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/MetadataFieldConverter.java
… dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/MetadataSchemaConverter.java