This sets the stage for projections support and implements "ListProjection", which is automatically applied by default when GETting resources that return lists.

It is expected that future PRs will a) improve efficiency by implementing true sub-resource paging via link repositories, while designating such linked sub-resources as optionally embeddable, and b) implement new projections we have not thought of yet.

With this branch, all HAL GET endpoints support an optional `?projection=name` parameter, which causes the projection, if found, to be applied before the result is returned. By default, endpoints that return lists of items are provided using the 'list' projection, which omits any embeds that are designated as optional.

I recommend reading the javadocs of `Projection` and `ConverterService` first, to get a general sense of how projections are expressed and used in this design.

**CURRENT STATE as of Nov 7:** 

* Main projections support is implemented. It allows creating projections that apply to objects at multiple points in their life cycle. 
* There is some significant refactoring, which effectively removes a lot of boilerplate code and consolidates JPAEntity -> RestModel conversions, and RestModel -> HALResource conversions into a common service usable anywhere, which accepts projections as input: `ConverterService`.
* To show what is possible, current code in this branch turns Item->Relationship embed into one provided by a link repository, and flags that as an optional embed. The idea is that when an embed is optional, the 'list' projection has the power to veto its being embedded (so it is just linked instead).
* To test, run on any repo that has relationships, then view the items/ endpoint that just lists items. Notice how for all items included, they all link to their relationships, they do not embed them. This is because the `ItemRestRepository` has told the `ConverterService` to automatically use the `ListProjection` when `findAll` is called.

Slides from Oct 17 WG meeting discussing what had been done up to that point: [Download PDF](https://wiki.duraspace.org/download/attachments/158269444/2019-10-17-Projections.pdf?version=1&modificationDate=1571325219372&api=v2)