@benbosman 2019-11-08
I did perform tests using this branch with an item containing 60 relationships. What I noticed is:
* The performance did improve (using 100 requests to the same item each time, always after already having requested the item at least 10 times prior to starting the test)
  * The item page using the master branch (containing relationships) required 210ms
  * The item page using this PR without specifying projections (containing relationships) required 433ms. No idea why it increased so much.
  * The item page using this PR with projection=list (excluding relationships) required 77ms. This is indeed a substantial improvement.
* When performing the same test on an item containing 1 relationship (but also less metadata), the results were:
  * The item page using the master branch (containing the relationship) required 27ms
  * The item page using this PR without specifying projections (containing the relationship) required 66ms. The time increased again
  * The item page using this PR with projection=list (excluding the relationship) required 31ms, close to the duration using the master branch
* The HAL link to relationships is missing without specifying projections. Although it may not be necessary because the relationships are embedded, this may be breaking some functionality. At this point in time, the master Angular UI has issues with this branch (doesn't render all relationships)
* The embedded paging details are broken (without specifying projections). They only mention 1 page, and 20 totalElements. Due to this, the next and last links are missing as well. These are all correct using the master code

I did not review the code yet

This is a good first step and has a substantial improvement for items with many relationships, but it worries me that the performance degrades when embedding, and the performance even degrades when removing the embedding if there's just one relationship