@cwilper 2019-11-10
@benbosman thanks for giving it a test. I have fixed the two functional regressions you reported. It sounds like we should have some coverage of those cases in the tests as well. I will add those and some additional tests to cover the new functionality introduced here.

Your performance report was concerning, so I did some detailed analysis using a profiler and found some room for improvement around the code that looks at the LinkRest annotations on rest class methods. By introducing some caching, I was able to improve things quite a bit there.

After merging master in and putting all the above changes in place, I ran a set of tests similar to yours, using curl to retrieve an item HAL resource 100 times. One with 60 relationships, and one with 1. Here are my results:

* Get one item with 60 rels 100 times:
  * master branch: 32 seconds
  * this branch, projection=default: 30 seconds
  * this branch, projection=list: 23 seconds
* Get one item iwth 1 rel 100 times:
  * master branch: 4.1 seconds
  * this branch, projection=default: 3.8 seconds
  * this branch, projection=list: 3.3 seconds

I should point out that were I really expect this branch to have a big effect is in endpoints that return pages of things (items, etc). So while looking at the effect on a single item is informative to zero in on differences between master, I expect the main benefit to be much more pronounced just by the fact that "projection=list" is automatically done on resources that are collections (this is akin to an "excerpt" projection as described in Spring Data Rest)