@benbosman 2019-11-12
Thanks @cwilper 

I also verified my tests using this branch with an item containing 60 relationships. What I noticed is:
* The performance did improve with the latest updates:
  * The item page using the master branch (containing relationships) required 210ms (not tested again)
  * The item page using this PR without specifying projections (containing relationships) required 191ms. This is indeed an improvement
  * The item page using this PR with projection=list (excluding relationships) required 73ms. Similar to the last test

When performing the same test on an item containing 1 relationship (but also less metadata), the results were:
  * The item page using the master branch (containing the relationship) required 27ms (not tested again)
  * The item page using this PR without specifying projections (containing the relationship) required 22ms. This is indeed an improvement
  * The item page using this PR with projection=list (excluding the relationship) required 15ms, an even bigger improvement.

The HAL link to relationships and the embedded paging details are resolved.

I will check out the code next, but the tests are looking very promising to me