@tdonohue 2019-11-21
@cwilper : If you can rebase this on the latest `master` and resolve the merge conflicts, I just merged #2529 and #2548 (as previously agreed upon).  Both of these add some code that will need to be updated to support Projections.

Once those updates are made to this PR, we can merge this (it's already at +2).